implemented initializer update-config hooks
