+++ src/Sulu/Bundle/AdminBundle/Resources/js/services/Initializer/types.js
@@ -1,0 +1,3 @@
+// @flow
+
+export type UpdateConfigHook = (config: Object, initialized: boolean) => void;
